refactor(tools): move brush/pencil/eraser stroke math out of paint-handlers (#451)
Extract per-tool stroke logic into co-located files:
- brush-stroke.ts: 130-line brush case with speed-size, jitter, taper,
  fade, scatter, and sub-brush dispatch
- pencil-stroke.ts: pencil line drawing with symmetry
- eraser-stroke.ts: eraser dab interpolation with symmetry

Combined with the earlier mask-handler collapse, paint-handlers.ts
drops from 898 to 692 lines. The switch in handlePaintMove is now a
thin 12-line dispatcher.
